Move challenge dialog type to its own file
