+++ stem-explorer-ng/src/app/shared/enums/challenge-dialog-type.enum.ts
@@ -1,0 +1,4 @@
+export enum ChallengeDialogType {
+  Preview,
+  Hint,
+}
